Moved citation html creation to a single function
